♻️ Update generators scripts

+++ scripts/mcunit.py
@@ -18,6 +18,8 @@ def create_symlink(src: Path, dest: Path):
 
 
 def clear_directory(dir: Path):
+    if not dir.is_dir():
+        return
     for entry in dir.iterdir():
         try:
             entry.unlink()
@@ -36,17 +38,31 @@ class MCUnit():
         Initializes MCUnit with the specified Minecraft version.
         """
         self.mc_version = mc_version
-        self.packtest_url = self.__get_modrinth_url("XsKUhp45")
-        self.fabric_api_url = self.__get_modrinth_url("P7dR8mSH")
-        self.fabric_server_url = f"https://meta.fabricmc.net/v2/versions/loader/{mc_version}/0.15.7/1.0.0/server/jar"
+        self.fabric_server_url = self.get_fabric_url(mc_version)
+        self.fabric_api_url = self.get_modrinth_url("P7dR8mSH", mc_version)
+        self.packtest_url = self.get_modrinth_url("XsKUhp45", mc_version)
 
-    def __get_modrinth_url(self, project_id: str) -> str:
+    def get_fabric_url(self, mc_version: str) -> str:
+        response = requests.get(f"https://meta.fabricmc.net/v2/versions/loader/{mc_version}")
+        response.raise_for_status()
+        versions = response.json()
+        return f'https://meta.fabricmc.net/v2/versions/loader/{mc_version}/{versions[0]["loader"]["version"]}/1.0.1/server/jar'
+
+    def get_modrinth_url(self, project_id: str, mc_version: str) -> str:
         response = requests.get(f"https://api.modrinth.com/v2/project/{project_id}/version")
         response.raise_for_status()
         versions = response.json()
-        if versions := [version for version in versions if self.mc_version in version["game_versions"]]:
+        if versions := [version for version in versions if mc_version in version["game_versions"]]:
             return versions[0]["files"][0]["url"]
-        raise RuntimeError(f"Could not find packtest version for the given Minecraft version: {self.mc_version}")
+        raise RuntimeError(f"Could not find packtest version for the given Minecraft version: {mc_version}")
+
+    def download_file(self, url: str, filepath: Path) -> None:
+        if not filepath.exists():
+            print(f"\033[93m🚀 Downloading\033[0m [{url}]")
+            response = requests.get(url)
+            response.raise_for_status()
+            filepath.parent.mkdir(parents=True, exist_ok=True)
+            filepath.write_bytes(response.content)
 
     def run(self, datapacks_path: str) -> None:
         """
@@ -62,12 +78,7 @@ class MCUnit():
             (self.fabric_api_url, tmpdir / "mods/fabric-api.jar"),
             (self.packtest_url, tmpdir / "mods/packtest.jar"),
         ]:
-            filepath.parent.mkdir(parents=True, exist_ok=True)
-            if not filepath.exists():
-                print(f"\033[93m🚀 Downloading\033[0m [{url}]")
-                response = requests.get(url)
-                response.raise_for_status()
-                filepath.write_bytes(response.content)
+            self.download_file(url, filepath)
 
         print("🔗 Linking datapacks")
         clear_directory(tmpdir / "world")
@@ -78,22 +89,22 @@ class MCUnit():
         (tmpdir / "allowed_symlinks.txt").write_text(str(datapacks_src_path))
 
         print("🧪 Running test server")
-        process = subprocess.Popen(
+        with subprocess.Popen(
             "java -Xmx2G -Dpacktest.auto -Dpacktest.auto.annotations -jar server.jar nogui",
             stdout=subprocess.PIPE,
             universal_newlines=True,
             shell=True,
             cwd=tmpdir,
-        )
+        ) as process:
+            for log in iter(process.stdout.readline, ""):
+                if match := re.search(r"::error title=(.*?)::(.*)", log):
+                    title, description = match.groups()
+                    print(f"\033[1m\033[91m ✘ {title}\033[0m: {description}")
 
-        for log in iter(process.stdout.readline, ""):
-            if match := re.search(r"::error title=(.*?)::(.*)", log):
-                title, description = match.groups()
-                print(f"\033[1m\033[91m ✘ {title}\033[0m: {description}")
-
-        if err_count := process.wait():
-            return print(f"💥 {err_count} required tests failed :(")
-        print("✅ All required tests passed :)")
+            if err_count := process.wait():
+                print(f"💥 {err_count} required tests failed :(")
+            else:
+                print("✅ All required tests passed :)")
 
 
 def run(datapacks_path: str, mc_version: str):
